+++ run.py
@@ -11,6 +11,8 @@ from flywheel_gear_toolkit import GearToolkitContext
 from app.command_line import exec_command
 from app.findMatchedScans import find_files
 from utils.niftiHeader import pixSize
+from utils.demo import get_demo
+
 # The gear is split up into 2 main components. The run.py file which is executed
 # when the container runs. The run.py file then imports the rest of the gear as a
 # module.
@@ -19,6 +21,9 @@ log = logging.getLogger(__name__)
 
 def main(context: GearToolkitContext) -> None:
     """Parses config and runs."""
+
+    subject_label, session_label = get_demo()
+
     # If one input is given no sub folders are created, so check if these exist, if not run find_files
     if not os.path.exists('/flywheel/v0/input/cor') or not os.path.exists('/flywheel/v0/input/sag'):
         find_files()
@@ -27,7 +32,7 @@ def main(context: GearToolkitContext) -> None:
     pixdim = pixSize()
 
     # Main event
-    command = "/flywheel/v0/app/ciso-gear.sh" + " " + str(pixdim)
+    command = "/flywheel/v0/app/ciso-gear.sh" + " " + subject_label # + str(pixdim)
     print(command)
     exec_command(
     command,